feat: multiple push providers
fix: fix naming

fix: more fixes

fix: more fixes

fix: add activity service

fix: update naming
